fix: resolve circular ESM import between frame.tsx and NeighborTiles.tsx
- Extract core DynamicFrame logic to DynamicFrameCore.tsx to break circular dependency
- Update frame.tsx to use render prop pattern for NeighborTiles injection
- Update NeighborTiles.tsx to use DynamicFrameCore directly
- Remove neighbor tiles opacity (now at full opacity)
- Add hierarchy panel transparency (50% opacity with backdrop blur)
- Apply canvas translation offset to loading skeleton for consistent positioning
- Remove overflow-hidden from canvas containers to enable transparency
