+++ src/app/map/Canvas/Tile/Item/content.tsx
@@ -33,7 +33,7 @@ export const DynamicTileContent = ({
   const title = data.title ?? 'Untitled';
   
   // Scale-based styling
-  const baseFontSize = scale === 1 ? "text-xs" : scale === 2 ? "text-base" : "text-lg";
+  const baseFontSize = scale === 1 ? "text-[11px]" : scale === 2 ? "text-base" : "text-lg";
   const fontWeight = scale === 1 ? "font-medium" : scale === 2 ? "font-medium" : "font-semibold";
   
   const testId = "tile-content";
