+++ src/app/map/layout.tsx
@@ -1,6 +1,6 @@
 export default function MapLayout({ children }: { children: React.ReactNode }) {
   return (
-    <main className="h-screen w-full overflow-hidden p-0 bg-background">
+    <main className="h-screen w-full p-0 bg-background">
       {children}
     </main>
   );
